NANO ESP32 METHOD WIP
This is untested as of this upload but compiled ok.  This is intended for a Arduino Nano ESP32 Specifically .  I would have to make special little wire manifolds for GND and VCC to share GPIO pins in order to use this correctly.  The GPIO Pins for the most part are all the same however...

The reason this version exists is to solve the LEDC Compilation errors - Solving these would give us better PWM control for smoother robot motion.  Breakout board is ordered from Amazon and on the way.  It is not as fun looking as the Freenove Breakout board however because there are no cool LEDs :D